Crate tidb_query_normal_expr

This crate implements a simple SQL query engine to work with TiDB pushed down executors.

The query engine is able to scan and understand rows stored by TiDB, run against a series of executors and then return the execution result. The query engine is provided via TiKV Coprocessor interface. However standalone UDF functions are also exported and can be used standalone.

Modules

builtin_arithmetic

builtin_cast

builtin_compare

builtin_control

builtin_encryption

builtin_json

builtin_like

builtin_math

builtin_miscellaneous

builtin_op

builtin_other

builtin_string

builtin_time

column

constant

scalar_function

Structs

Column

Constant

ScalarFunc
A single scalar function call

Enums

Error

Expression

Functions

eval_arith

Type Definitions

Result